Fix compiling errors ( updateValue had to be a pure virtual method )

+++ CPP_GeneticAlgorithm_Army/Cap_Health.h
@@ -11,7 +11,7 @@ public:
 	}
 
 	void updateValue(){
-		_maxLife = (1 + _level) * 10;
+		_maxLife = (float)((1 + _level) * 10);
 	}
 
 	// add method to update _value
